Changelog update, include require_relative for invoice/rule, refactor InvoiceRule to Invoice::Rule
