+++ lib/invoice/invoice.rb
@@ -24,7 +24,7 @@ module StarkBank
   # - fine [float, default 0.0]: Invoice fine for overdue payment in %. ex: 2.5
   # - interest [float, default 0.0]: Invoice monthly interest for overdue payment in %. ex: 5.2
   # - discounts [list of hashes, default nil]: list of hashes with 'percentage':float and 'due':DateTime or string pairs
-  # - @TODO: default [] or nil? rules [list of InvoiceRule, default []]: list of Invoice.Rule objects for modifying invoice behavior. ex: [Invoice.Rule(key="allowedTaxIds", value=[ "012.345.678-90", "45.059.493/0001-73" ])]
+  # - rules [list of Invoice::Rule, default []]: list of Invoice.Rule objects for modifying invoice behavior. ex: [Invoice.Rule(key="allowedTaxIds", value=[ "012.345.678-90", "45.059.493/0001-73" ])]
   # - descriptions [list of hashes, default nil]: list of hashes with 'key':string and 'value':string pairs
   # - tags [list of strings, default nil]: list of strings for tagging
   #
@@ -61,7 +61,7 @@ module StarkBank
       @pdf = pdf
       @link = link
       @descriptions = descriptions
-      @rules = rules
+      @rules = StarkBank::Invoice::Rule.parse_rules(rules)
       @nominal_amount = nominal_amount
       @fine_amount = fine_amount
       @interest_amount = interest_amount
