+++ lib/starkbank.rb
@@ -5,6 +5,7 @@ require_relative('balance/balance')
 require_relative('transaction/transaction')
 require_relative('invoice/invoice')
 require_relative('invoice/log')
+require_relative('invoice/rule')
 require_relative('invoice/payment')
 require_relative('dict_key/dict_key')
 require_relative('dynamic_brcode/dynamic_brcode')
